softlink directory
I need to install the soft links in a custom directory in CMAKE_INSTALL_PREFIX. What do you propose is the best way to do this? This PR contains untested suggestion. The goal is for a library to install softlinks to files in /opt/rocm/$library/include* to /opt/rocm/include/$library. Now by default the softlinks go to /opt/rocm/include/
If there is another way to do it, that will be great.