ROCMSetupVersion.cmake: Remove "dirty" keyword from git command
Since stage2 builds in the CI environment are run in docker
using docker build command, the source code is copied over
from the BuildSlave to the docker container. Git log detects
this and adds dirty keyword to the package names.
Remove this keyword until we find the solution to swithc over
from docker build to docker run while mounting the mathlibs or
in the container figure out the way to recreate HEAD.

Change-Id: Ic1d9100ff78e7bc578c0828a3866d2e5fcbc75bd
Signed-off-by: Aakash Sudhanwa <Aakash.Sudhanwa@amd.com>