Make rocm_check_target_ids more robust (#75)
* Make rocm_check_target_ids more robust

Add support for target features for rocm_check_target ids. Before this
change, targets with target features specified (like xnack or sramecc)
would fail when -Werror was set.

* Address code review comments and fix linter errors

* Improve target id variable name

The old HAVE_<target-id> was a bit unclear about what component it was
checking for the GPU architecture. Since we're changing the variable
name to fix the behaviour with -Werror, this is a good opportunity to
make the name more explicit at the same time.

* Tweak _rocm_sanitize_target_id

Use lowercase 'on' and 'off' as all target ids are currently lowercase
and it looks less jarring.

* Update CHANGELOG.md

Co-authored-by: Cordell Bloor <Cordell.Bloor@amd.com>