Fix use of LIBDIR
All of the math libs seem to have this line:

set(CMAKE_INSTALL_LIBDIR "lib" CACHE INTERNAL "..." FORCE)

To ensure that libdir is always predicably "lib". The problem is that
force is used, so it cannot be overriden. After some digging, it looks
like all of them include ROCMInstallTargets, which includes
GNUInstallerDirs.

While GNUInstallerDirs is a good idea and is recommended, it has some
logic to autodetect the libdir if unset, which we don't want. It seems
that each library is forcing the value of this because this
functionality of GNUInstallerDirs is not commonly known.

The solution is just to set a default LIBDIR prior to including
GNUInstallerDirs, which is done with other ROCm libs, such as OpenCL and
ROCr. Once done, the force set of LIBDIR can be dropped from other
components that use rocm-cmake.

I also set ROCM_INSTALL_LIBDIR to CMAKE_INSTALL_LIBDIR to make sure
these variables are the same value, as it's typical for users or distros
to override the publically documented CMAKE_INSTALL_LIBDIR:

https://cmake.org/cmake/help/latest/module/GNUInstallDirs.html

Signed-off-by: Jeremy Newton <Jeremy.Newton@amd.com>