Set project languages to NONE
Set the languages of the projects to NONE explicitly, to avoid CMake defaulting to C and C++.

From [CMake command project()](https://cmake.org/cmake/help/latest/command/project.html?highlight=project):
> By default C and CXX are enabled if no language options are given

This speeds up the configuration by a small amount because cmake does not have to search for and verify the compilers.
It also allows configuring the project when a C orC++ compilers are not available. Finally when the build of rocm-cmake
is embedded in an other project (as in the case of rocRAND) it avoids errors caused by cmake selecting MSVC
for the C compiler but Clang for the C++ compiler on windows.