Fix version-norepo and version-noparent tests (#117)
This changes the tests to only match the numeric component, like the 
implementation of `rocm_setup_version`. Prior to these changes, the 
actual value would be the commit hash but the expected value would be
the long output of git describe (containing the name of the nearest
tag, the number of commits since that tag, and the commit hash). i.e.,

expected: rocm-5.3.3-7-g8bc91a1
actual: g8bc91a1

This changes it to be:

expected: g8bc91a1
actual: g8bc91a1

I still don't understand why this only recently began failing on the CI
and I don't have a clear understanding of what these tests are testing.