make Clang-Tidy and CppCheck optional (#135)
* make clang-tidy optional if the tool is not found

add the dependent option to let clang-tidy check off even when
the tool is found. The reason is a vast number of targets
generated which pollutes the project view when opened in IDE
(i.e., Microsoft Visual Studio, JetBrains Clion).

* make cppcheck optional if the tool is not found

add the dependent option to let cppcheck off even when the
tool is found. The reason is a vast number of targets
generated which pollutes the project view when opened in IDE
(i.e., Microsoft Visual Studio, JetBrains Clion).

---------

Co-authored-by: Artur Wojcik <artur.wojcik@amd.com>