Rename CMake package to ROCmCMakeBuildTools (#151)
For discussion, see: Rename CMake package to ROCmCMakeBuildTools (#128).

Having the name of this CMake package be ROCM is consistently confusing to users who expect find_package(ROCM [...]) to add ROCm packages to their build, instead of tools that we use to build ROCm (see #49). Renaming the CMake project to ROCmCMakeBuildTools makes the purpose of this package clearer.

In order to have as smooth a transition as possible, starting this as a notice, will make into a deprecation/warning in ~6.2, error sometime later, remove in 7.0.